Move to the beginning of buffer once narrowed.
Instead of searching for the beginning of the current section by
matching a star at the beginning of a line, we just move to the
beginning of the buffer. Narrowing ensure that we are really moving to
the beginning of the current section.

The previous solution was not working if the point was already at the
beginning of the current section.